directory changes, and added reith font to the project
