+++ apti/utilities.py
